Commited exercise4-21 (Blogs with users)
Added authorization checks for deleting a blog

+++ controllers/blogs.js
@@ -40,6 +40,13 @@ blogRouter.post("/", async (request, response) => {
 })
 
 blogRouter.delete("/:id", async (request, response) => {
+    const token = tokenizer.verify(request.token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id)
+    if (token.id !== blog.user.toString()) {
+        return response.status(401).json({
+            error: "Unauthorized - the blog could not be deleted"
+        })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
@@ -47,7 +54,6 @@ blogRouter.delete("/:id", async (request, response) => {
 blogRouter.put("/:id", async (request, response) => {
     const body = request.body
     const originalValues = await Blog.findById(request.params.id)
-    console.log(originalValues)
     const blog = {
         title: body.title || originalValues.title,
         author: body.author || originalValues.author,
